Extract ParserUtils class from Parser class
