+++ src/main/java/seedu/splitlah/command/ActivityListCommand.java
@@ -6,6 +6,7 @@ import seedu.splitlah.data.Session;
 import seedu.splitlah.exceptions.InvalidDataException;
 import seedu.splitlah.exceptions.InvalidFormatException;
 import seedu.splitlah.parser.Parser;
+import seedu.splitlah.parser.ParserUtils;
 import seedu.splitlah.ui.Message;
 
 import java.util.ArrayList;
@@ -21,7 +22,7 @@ public class ActivityListCommand extends Command {
     public static final String COMMAND_FORMAT = "Syntax: activity /list /sid [SESSION_ID]";
 
     public static final String[] COMMAND_DELIMITERS = {
-        Parser.SESSION_ID_DELIMITER
+        ParserUtils.SESSION_ID_DELIMITER
     };
 
     private int sessionId;
