+++ src/main/java/seedu/splitlah/command/ActivityViewCommand.java
@@ -6,6 +6,7 @@ import seedu.splitlah.data.Session;
 import seedu.splitlah.exceptions.InvalidDataException;
 import seedu.splitlah.exceptions.InvalidFormatException;
 import seedu.splitlah.parser.Parser;
+import seedu.splitlah.parser.ParserUtils;
 
 
 /**
@@ -20,8 +21,8 @@ public class ActivityViewCommand extends Command {
     public static final String COMMAND_FORMAT = "Syntax: activity /view /sid [SESSION_ID] /aid [ACTIVITY_ID]";
 
     public static final String[] COMMAND_DELIMITERS = { 
-        Parser.SESSION_ID_DELIMITER,
-        Parser.ACTIVITY_ID_DELIMITER
+        ParserUtils.SESSION_ID_DELIMITER,
+        ParserUtils.ACTIVITY_ID_DELIMITER
     };
 
     private int sessionId;
