+++ src/main/java/seedu/splitlah/command/SessionDeleteCommand.java
@@ -5,6 +5,7 @@ import seedu.splitlah.data.Session;
 import seedu.splitlah.exceptions.InvalidDataException;
 import seedu.splitlah.exceptions.InvalidFormatException;
 import seedu.splitlah.parser.Parser;
+import seedu.splitlah.parser.ParserUtils;
 import seedu.splitlah.ui.Message;
 
 import java.util.logging.Level;
@@ -24,7 +25,7 @@ public class SessionDeleteCommand extends Command {
             "The session was deleted successfully.";
 
     public static final String[] COMMAND_DELIMITERS = {
-        Parser.SESSION_ID_DELIMITER
+        ParserUtils.SESSION_ID_DELIMITER
     };
 
     private int sessionId;
