+++ src/main/java/seedu/splitlah/data/Session.java
@@ -1,11 +1,10 @@
 package seedu.splitlah.data;
 
 import seedu.splitlah.exceptions.InvalidDataException;
-import seedu.splitlah.parser.Parser;
+import seedu.splitlah.parser.ParserUtils;
 import seedu.splitlah.ui.Message;
 
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 
 /**
@@ -346,7 +345,7 @@ public class Session {
      */
     public String getSessionSimplifiedString() {
         return sessionId + SUMMARY_STRING_SEPARATOR + sessionName + "\n "
-                + SUMMARY_STRING_SEPARATOR + dateCreated.format(Parser.DATE_FORMAT)
+                + SUMMARY_STRING_SEPARATOR + dateCreated.format(ParserUtils.DATE_FORMAT)
                 + SUMMARY_STRING_SEPARATOR + personList.size() + " participants"
                 + SUMMARY_STRING_SEPARATOR + activityList.size() + " activities"
                 + SUMMARY_STRING_SEPARATOR + getGroupSummaryString();
@@ -361,7 +360,7 @@ public class Session {
     public String toString() {
         return "Session Id #" + sessionId + " --\n"
                 + "Name: " + sessionName + '\n'
-                + "Date: " + dateCreated.format(Parser.DATE_FORMAT) + '\n'
+                + "Date: " + dateCreated.format(ParserUtils.DATE_FORMAT) + '\n'
                 + getActivityListSummaryString() + '\n'
                 + getPersonListSummaryString() + '\n'
                 + getGroupSummaryString();
